Move MetroBarChart to components/index/cards/charts
