+++ components/index/cards/charts/MetroBarChart.vue
